Refactor landmarking scripts into separate folder - Updated tests - Updated documentation - Fixed typos in docs - Set llvm version to 3.3 for Travis
