+++ morphman/misc/__init__.py
@@ -4,6 +4,3 @@ from .estimate_alpha_and_beta import estimate_alpha_and_beta, compute_quantities
                                      odr_line, get_moved_siphon, find_angle, \
                                      find_angle_odr, write_alpha_beta_point, \
                                      alpha_beta_intersection
-from .automated_landmarking import automated_landmarking, landmarking_bogunovic, \
-                                   landmarking_piccinelli, map_landmarks, \
-                                   spline_centerline_and_compute_geometric_features, create_particles
